requirements.txt
certifi==2019.9.11
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.3.1
cycler==0.10.0
idna==2.8
joblib==0.14.0
kiwisolver==1.1.0
matplotlib==3.1.1
MedPy==0.4.0
nose==1.3.7
nose-timer==0.7.5
numpy==1.17.3
opencv-python>=4.1.1,<4.2
Pillow==6.1.0
pycparser==2.20
pycryptodome==3.9.9
pycryptodomex==3.9.9
#pygame==1.9.6
pyOpenSSL==20.0.1
pyparsing==2.4.2
PyQt5==5.15.2
pyqt5-plugins==5.15.2.2.0.1
PyQt5-sip==12.8.1
pyqt5-tools==5.15.2.3
pyuic5-tool==0.0.1
python-dateutil==2.8.0
python-dotenv==0.15.0
python-resize-image==1.1.19
qimage2ndarray
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
requests==2.22.0
#scikit-learn==0.21.3
#scipy==1.3.1
SimpleITK==2.0.2
six==1.12.0
sklearn==0.0
urllib3==1.25.6
